refactor: colocate style utilities with related components
Move shared style utilities from ui/styles/ to their related component folders:
- fields.ts → ui/common/forms/styles.ts
- interactive.ts → ui/common/buttons/styles.ts
